Concurrency_7 int changed to long

+++ Concurrency_7/src/main/java/ru/nsu/fit/markelov/Department.java
@@ -1,20 +1,20 @@
 package ru.nsu.fit.markelov;
 
 public class Department {
-    private final int threadId;
-    private final int threadsCount;
-    private final int iterations;
+    private final long threadId;
+    private final long threadsCount;
+    private final long iterations;
 
     private double calculationResult = 0;
 
-    public Department(int threadId, int threadsCount, int iterations) {
+    public Department(long threadId, long threadsCount, long iterations) {
         this.threadId = threadId;
         this.threadsCount = threadsCount;
         this.iterations = iterations;
     }
 
     public void performCalculations() {
-        for (int i = 0; i < iterations; i++) {
+        for (long i = 0; i < iterations; i++) {
             long n = i * threadsCount + threadId; // number of the element in the Leibniz formula
             double nthElementValue = 1d / (2 * n + 1);
